feat(tenant): my tenant member manager page

+++ src/main/kotlin/com/lovelycatv/crystalframework/tenant/controller/manager/member/ManagerTenantMemberController.kt
@@ -2,15 +2,18 @@ package com.lovelycatv.crystalframework.tenant.controller.manager.member
 
 import com.lovelycatv.crystalframework.rbac.constants.SystemPermission
 import com.lovelycatv.crystalframework.shared.constants.GlobalConstants
+import com.lovelycatv.crystalframework.shared.exception.ForbiddenException
+import com.lovelycatv.crystalframework.shared.exception.UnauthorizedException
 import com.lovelycatv.crystalframework.shared.response.ApiResponse
 import com.lovelycatv.crystalframework.shared.types.UserAuthentication
+import com.lovelycatv.crystalframework.shared.utils.RbacUtils
+import com.lovelycatv.crystalframework.tenant.constants.TenantPermission
 import com.lovelycatv.crystalframework.tenant.controller.manager.member.dto.ManagerCreateTenantMemberDTO
 import com.lovelycatv.crystalframework.tenant.controller.manager.member.dto.ManagerDeleteTenantMemberDTO
 import com.lovelycatv.crystalframework.tenant.controller.manager.member.dto.ManagerReadTenantMemberDTO
 import com.lovelycatv.crystalframework.tenant.controller.manager.member.dto.ManagerUpdateTenantMemberDTO
 import com.lovelycatv.crystalframework.tenant.service.manager.TenantMemberManagerService
 import jakarta.validation.Valid
-import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
 
@@ -20,7 +23,6 @@ import org.springframework.web.bind.annotation.*
 class ManagerTenantMemberController(
     private val tenantMemberManagerService: TenantMemberManagerService
 ) {
-    @PreAuthorize("hasAnyAuthority('${SystemPermission.ACTION_TENANT_MEMBER_CREATE}')")
     @PostMapping("/create", version = "1")
     suspend fun create(
         userAuthentication: UserAuthentication,
@@ -28,11 +30,15 @@ class ManagerTenantMemberController(
         @Valid
         dto: ManagerCreateTenantMemberDTO
     ): ApiResponse<*> {
-        tenantMemberManagerService.create(dto)
+        if (RbacUtils.hasAuthority(SystemPermission.ACTION_TENANT_MEMBER_CREATE)) {
+            tenantMemberManagerService.create(dto)
+        } else {
+            throw ForbiddenException()
+        }
+
         return ApiResponse.success(null)
     }
 
-    @PreAuthorize("hasAnyAuthority('${SystemPermission.ACTION_TENANT_MEMBER_READ}')")
     @GetMapping("/query", version = "1")
     suspend fun query(
         userAuthentication: UserAuthentication,
@@ -40,10 +46,21 @@ class ManagerTenantMemberController(
         @Valid
         dto: ManagerReadTenantMemberDTO
     ): ApiResponse<*> {
-        return ApiResponse.success(tenantMemberManagerService.queryVO(dto))
+        val result = if (RbacUtils.hasAuthority(SystemPermission.ACTION_TENANT_MEMBER_READ)) {
+            tenantMemberManagerService.queryVO(dto)
+        } else if (RbacUtils.hasAuthority(TenantPermission.ACTION_TENANT_MEMBER_READ_PEM)) {
+            if (dto.tenantId == userAuthentication.tenantId) {
+                tenantMemberManagerService.queryVO(dto)
+            } else {
+                throw UnauthorizedException()
+            }
+        } else {
+            throw ForbiddenException()
+        }
+
+        return ApiResponse.success(result)
     }
 
-    @PreAuthorize("hasAnyAuthority('${SystemPermission.ACTION_TENANT_MEMBER_UPDATE}')")
     @PostMapping("/update", version = "1")
     suspend fun update(
         userAuthentication: UserAuthentication,
@@ -51,11 +68,22 @@ class ManagerTenantMemberController(
         @Valid
         dto: ManagerUpdateTenantMemberDTO
     ): ApiResponse<*> {
-        tenantMemberManagerService.update(dto)
+        if (RbacUtils.hasAuthority(SystemPermission.ACTION_TENANT_MEMBER_UPDATE)) {
+            tenantMemberManagerService.update(dto)
+        } else if (RbacUtils.hasAuthority(TenantPermission.ACTION_TENANT_MEMBER_UPDATE_PEM)) {
+            userAuthentication.assertTenantIdNotNull()
+            if (tenantMemberManagerService.checkIsRelated(dto.id, userAuthentication.tenantId!!)) {
+                tenantMemberManagerService.update(dto)
+            } else {
+                throw UnauthorizedException()
+            }
+        } else {
+            throw ForbiddenException()
+        }
+
         return ApiResponse.success(null)
     }
 
-    @PreAuthorize("hasAnyAuthority('${SystemPermission.ACTION_TENANT_MEMBER_DELETE}')")
     @PostMapping("/delete", version = "1")
     suspend fun delete(
         userAuthentication: UserAuthentication,
@@ -63,7 +91,19 @@ class ManagerTenantMemberController(
         @Valid
         dto: ManagerDeleteTenantMemberDTO
     ): ApiResponse<*> {
-        tenantMemberManagerService.deleteByDTO(dto)
+        if (RbacUtils.hasAuthority(SystemPermission.ACTION_TENANT_MEMBER_DELETE)) {
+            tenantMemberManagerService.deleteByDTO(dto)
+        } else if (RbacUtils.hasAuthority(TenantPermission.ACTION_TENANT_MEMBER_DELETE_PEM)) {
+            userAuthentication.assertTenantIdNotNull()
+            if (tenantMemberManagerService.checkIsRelated(dto.ids, userAuthentication.tenantId!!)) {
+                tenantMemberManagerService.deleteByDTO(dto)
+            } else {
+                throw UnauthorizedException()
+            }
+        } else {
+            throw ForbiddenException()
+        }
+
         return ApiResponse.success(null)
     }
 }
